+++ lib/src/metrics_analyzer_utils.dart
@@ -1,5 +1,6 @@
 import 'package:analyzer/dart/ast/ast.dart';
-import 'package:dart_code_metrics/src/cyclomatic_complexity/models/scoped_declaration.dart';
+
+import 'models/scoped_declaration.dart';
 
 int getArgumentsCount(ScopedDeclaration dec) {
   final declaration = dec.declaration;
@@ -15,7 +16,7 @@ int getArgumentsCount(ScopedDeclaration dec) {
   return argumentsCount ?? 0;
 }
 
-String getQualifiedName(ScopedDeclaration dec) {
+String getHumanReadableName(ScopedDeclaration dec) {
   final declaration = dec.declaration;
 
   if (declaration is FunctionDeclaration) {